Added load config because interacting with CLI is annoying.
Added load keywords because waiting for API calls is annoying.

Converted Global_state.llmOutput to be a map of product_id to llmOutput
because working with deques/arrays is annoying.

Converted Global_state.reviews to be a map of product_id to review chunks
because working with arrays of arrays is annoying.

Added in Zeru's stemmer changes.

Begin integrating updating global state with Ketan's
updated aggregator code.